This is for unconfined login user subjects using xserver.
Implement a set of dontaudit base catchall type attributes.
This is for application/shell subjects that need to be silenced forcefully.

Consider for example running an app from some exotic location. The app may want to traverse it, but this should not be strictly required.
If you would add allow rules for all possibilities you'd be adding a lot of rules potentially.

Renamed sysctl_t to sysctl_fs_t for consistency.