I am pretty sure that one needs about the same permissions that one needs if one runs systemctl when one runs askpwdagent.
I need to test that with a user that is not a systemctl client subject to see (e.g. try to shutdown as unpriv user that has polkit allowance)

I removed some rules from systemctl client subject since common user subject already have those privs and i want to base it on that. (i moved the rules to systemctl_t)