Wishful thinking:
You should be able to create a base policy with systemd and just a few extras like journald, udev and a couple of dependencies.
You should not need the usersubject module in that scenario (think some embedded configuration)

I move nsswitch_config_t out of auth and into an ns module, i also moved host.conf out of miscfiles net conf and into ns.
This might make some calls to miscfiles_read_net_config_files no longer needed to i removed most of them to make sure.

I moved utmp and wtmp out of auth and into an utmp module. this module does not depend on systemd but it does support the systemd-utmp util if systemd is present.
The idea is that all init systems need utmp and wtmp.

I moved initctl out of systemd and made that more generic as well, Again if systemd is available then systemd-initctl is supported.
if not then other init systems can still use it for their initctl fifo

I added a new test for a minimal base policy plus just systemd and its strict dependencies. (i probably dont have this all correct but it is something)

I made auth module not depend on usersubject. login programs are for users. however if you do not have a login program that the usersubject rules
for login programs arent needed either obviously. So the dirty hack here was to make usersubject macro calls for login programs optional to make it build.
It does not look right but there is a compelling argument for this decision i believe.

All in all this is a pretty nastly patch...