A lot of nastyness related to dbus sessions.
Theres a LOT of compromize in here for the sake of keeping it maintainable.

For example session clients rely on dac.

All application subject types can send child terminated signals to all systemd session subjects.
This is because applications may not be a direct child of the systemd session... e.g. a decendant of a systemd session app.

systemduser also needs to support socket activation

Again compromise: all systemduser subjects can deal with all systemd socket activated subjects.
socket activation for session daemon cannot quite work the same because the sockets are generally in tmpfs