Weird ...
Eventhough mplayer is not a dbususer client, it still was not able to get sound working with pa when pa was just a dbususer client subject
I suppose pa was not able to do something that it needed (worked in permissive mode): no avc denials whatsoever....
mplayer is able to get sound working fine with pa if pa does not get access to dbus at all.

Anyhow this commit makes mplayer sound with pa work (for some reason) when pa has access to dbus session bus...

Had to give it access to all busses though since pa_t is not a prefixed type.