This does not make sense
I was getting this after i added an unconfined user. I must have been running as root in that unconfined domain.

If you create new privileged user domains, and you want to use ubac, then you would need to exclude ubac.cil and include
a modified one in sources.d/ that included the identity of your privileged user.

This is caused by an limitation in the selinux policy language. We need userattribute to make this work.