Don't let usersubject_common_login_subject_type call these
It is too generic. I want the flexibility to exclude elements of the "user_object_template"

For example exclude tmpfs access, pty or tty access

It should be possible to exclude tmpfs access, even with systemduser.

It is systemd that maintains /run/user/UID/systemd, if you have a user that is not allowed to interact with its session daemon then that user does not need access to that location.